Replaced forward declaration of SolitaireGame with full one & moved most constants in solitaire.h to solitaire.cc. Replaced assert with SPIEL_CHECK_EQ in solitaire_test.cc
